refactor: centraliza funções utilitárias em src/lib/utils.js
Extrai initials(), fmtCurrency() e hashId() de 6 arquivos para uma
fonte canônica. Corrige divergência em fmtCurrency: !n tratava 0 como
inválido em Dashboard e EstrategiaModule; n == null não trata.

Remove export morto de fmtCurrency no ROICalculator (não importado
por nenhum arquivo externo).

Closes #48
