+++ src/components/ROICalculator.jsx
@@ -2,17 +2,12 @@ import PropTypes from 'prop-types'
 import { useState, useMemo, useCallback } from 'react'
 import { Calculator, Target, DollarSign, Save, BarChart3, TrendingUp, AlertCircle, FileDown } from 'lucide-react'
 import { exportROIPDF } from '../utils/exportPDF'
+import { fmtCurrency } from '../lib/utils'
 
 // ─── Formatters ───────────────────────────────────────────────────────────────
-export function fmt(n) {
-  if (n === null || n === undefined || isNaN(n)) return '—'
-  if (!isFinite(n)) return '∞'
+function fmt(n) {
+  if (n == null || isNaN(n) || !isFinite(n)) return '—'
   return n.toLocaleString('pt-BR', { maximumFractionDigits: 0 })
-}
-export function fmtCurrency(n) {
-  if (n === null || n === undefined || isNaN(n)) return '—'
-  if (!isFinite(n)) return '∞'
-  return n.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL', maximumFractionDigits: 0 })
 }
 
 // ─── Number input ─────────────────────────────────────────────────────────────
