+++ src/lib/utils.js
@@ -1,0 +1,18 @@
+// Funções utilitárias puras compartilhadas entre páginas e componentes
+
+export function initials(name = '') {
+  return name.split(' ').slice(0, 2).map(w => w[0]?.toUpperCase() || '').join('')
+}
+
+// n == null cobre tanto null quanto undefined sem tratar 0 como inválido
+export function fmtCurrency(n) {
+  if (n == null || isNaN(n) || !isFinite(n)) return '—'
+  return n.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL', maximumFractionDigits: 0 })
+}
+
+// Compatível com IDs numéricos legados e UUIDs
+export function hashId(id) {
+  if (!id) return 0
+  const n = Number(id)
+  return !isNaN(n) && isFinite(n) ? n : String(id).split('').reduce((a, c) => a + c.charCodeAt(0), 0)
+}
